+++ e2e/tests/upstreams.crud-all-fields.spec.ts
@@ -15,11 +15,15 @@
  * limitations under the License.
  */
 import { upstreamsPom } from '@e2e/pom/upstreams';
-import { genTLS, randomId } from '@e2e/utils/common';
+import { randomId } from '@e2e/utils/common';
 import { e2eReq } from '@e2e/utils/req';
 import { test } from '@e2e/utils/test';
 import { uiHasToastMsg } from '@e2e/utils/ui';
-import { expect, type Locator } from '@playwright/test';
+import {
+  uiCheckUpstreamAllFields,
+  uiFillUpstreamAllFields,
+} from '@e2e/utils/ui/upstreams';
+import { expect } from '@playwright/test';
 
 import { deleteAllUpstreams } from '@/apis/upstreams';
 
@@ -29,13 +33,6 @@ test.beforeAll(async () => {
 
 test('should CRUD upstream with all fields', async ({ page }) => {
   test.setTimeout(30000);
-
-  const fillHTTPStatuses = async (input: Locator, ...statuses: string[]) => {
-    for (const status of statuses) {
-      await input.fill(status);
-      await input.press('Enter');
-    }
-  };
 
   const upstreamNameWithAllFields = randomId('test-upstream-full');
   const description =
@@ -49,220 +46,9 @@ test('should CRUD upstream with all fields', async ({ page }) => {
   await upstreamsPom.getAddUpstreamBtn(page).click();
   await upstreamsPom.isAddPage(page);
 
-  await test.step('fill in required fields', async () => {
-    // Fill in the required fields
-    // 1. Name (required)
-    await page
-      .getByLabel('Name', { exact: true })
-      .fill(upstreamNameWithAllFields);
-
-    // 2. Description (optional but simple)
-    await page.getByLabel('Description').fill(description);
-
-    // 3. Add multiple nodes (required)
-    const addNodeBtn = page.getByRole('button', { name: 'Add a Node' });
-    const nodesSection = page.getByRole('group', { name: 'Nodes' });
-
-    // Wait for 'No Data' text to be visible
-    const noData = nodesSection.getByText('No Data');
-    await expect(noData).toBeVisible();
-
-    // Add the first node, using force option
-    await addNodeBtn.click();
-    await expect(noData).toBeHidden();
-
-    // Wait for table rows to appear
-    const rows = nodesSection.locator('tr.ant-table-row');
-    await expect(rows.first()).toBeVisible();
-
-    // Fill in the Host for the first node - click first then fill
-    const hostInput = rows.first().locator('input').first();
-    await hostInput.click();
-    await hostInput.fill('node1.example.com');
-    await expect(hostInput).toHaveValue('node1.example.com');
-
-    // Fill in the Port for the first node - click first then fill
-    const portInput = rows.first().locator('input').nth(1);
-    await portInput.click();
-    await portInput.fill('8080');
-    await expect(portInput).toHaveValue('8080');
-
-    // Fill in the Weight for the first node - click first then fill
-    const weightInput = rows.first().locator('input').nth(2);
-    await weightInput.click();
-    await weightInput.fill('10');
-    await expect(weightInput).toHaveValue('10');
-
-    // Fill in the Priority for the first node - click first then fill
-    const priorityInput = rows.first().locator('input').nth(3);
-    await priorityInput.click();
-    await priorityInput.fill('1');
-
-    // Add the second node with a more reliable approach
-    await nodesSection.click();
-    await addNodeBtn.click();
-
-    await expect(rows.nth(1)).toBeVisible();
-
-    // Fill in the Host for the second node - click first then fill
-    const hostInput2 = rows.nth(1).locator('input').first();
-    await hostInput2.click();
-    await hostInput2.fill('node2.example.com');
-    await expect(hostInput2).toHaveValue('node2.example.com');
-
-    // Fill in the Port for the second node - click first then fill
-    const portInput2 = rows.nth(1).locator('input').nth(1);
-    await portInput2.click();
-    await portInput2.fill('8081');
-    await expect(portInput2).toHaveValue('8081');
-
-    // Fill in the Weight for the second node - click first then fill
-    const weightInput2 = rows.nth(1).locator('input').nth(2);
-    await weightInput2.click();
-    await weightInput2.fill('5');
-    await expect(weightInput2).toHaveValue('5');
-
-    // Fill in the Priority for the second node - click first then fill
-    const priorityInput2 = rows.nth(1).locator('input').nth(3);
-    await priorityInput2.click();
-    await priorityInput2.fill('2');
-    await expect(priorityInput2).toHaveValue('2');
-  });
-
-  await test.step('fill in all optional fields', async () => {
-    // Fill in all optional fields
-
-    // 1. Load balancing type - using force option
-    await page
-      .getByRole('textbox', { name: 'Type', exact: true })
-      .scrollIntoViewIfNeeded();
-    await page.getByRole('textbox', { name: 'Type', exact: true }).click();
-    await page.getByRole('option', { name: 'chash' }).click();
-
-    // 2. Hash On field (only useful when type is chash) - using force option
-    await page.getByRole('textbox', { name: 'Hash On' }).click();
-    await page.getByRole('option', { name: 'header' }).click();
-
-    // 3. Key field (only useful when type is chash)
-    await page
-      .getByRole('textbox', { name: 'Key', exact: true })
-      .fill('X-Custom-Header');
-
-    // 4. Set protocol (Scheme) - using force option
-    await page.getByRole('textbox', { name: 'Scheme' }).click();
-    await page.getByRole('option', { name: 'https' }).click();
-
-    // 5. Set retry count (Retries)
-    await page.getByLabel('Retries').fill('5');
-
-    // 6. Set retry timeout (Retry Timeout)
-    await page.getByLabel('Retry Timeout').fill('6');
-
-    // 7. Pass Host setting - using force option
-    await page.getByRole('textbox', { name: 'Pass Host' }).click();
-    await page.getByRole('option', { name: 'rewrite' }).click();
-
-    // 8. Upstream Host
-    await page.getByLabel('Upstream Host').fill('custom.upstream.host');
-
-    // 9. Timeout settings
-    const timeoutSection = page.getByRole('group', { name: 'Timeout' });
-    await timeoutSection.getByLabel('Connect').fill('3');
-    await timeoutSection.getByLabel('Send').fill('3');
-    await timeoutSection.getByLabel('Read').fill('3');
-
-    // 10. Keepalive Pool settings
-    const keepaliveSection = page.getByRole('group', {
-      name: 'Keepalive Pool',
-    });
-    await keepaliveSection.getByLabel('Size').fill('320');
-    await keepaliveSection.getByLabel('Idle Timeout').fill('60');
-    await keepaliveSection.getByLabel('Requests').fill('1000');
-
-    // 11. TLS client verification settings
-    const tlsSection = page.getByRole('group', { name: 'TLS' });
-    const tls = genTLS();
-    await tlsSection
-      .getByRole('textbox', { name: 'Client Cert', exact: true })
-      .fill(tls.cert);
-    await tlsSection
-      .getByRole('textbox', { name: 'Client Key', exact: true })
-      .fill(tls.key);
-    await tlsSection
-      .locator('label')
-      .filter({ hasText: 'Verify' })
-      .locator('div')
-      .first()
-      .click();
-
-    // 12. Health Check settings
-    // Activate active health check
-    const healthCheckSection = page.getByRole('group', {
-      name: 'Health Check',
-    });
-    const checksEnabled = page.getByTestId('checksEnabled').locator('..');
-    await checksEnabled.click();
-
-    // Set the Healthy part of Active health check settings
-    const activeSection = healthCheckSection.getByRole('group', {
-      name: 'Active',
-    });
-    await activeSection
-      .getByRole('textbox', { name: 'Type', exact: true })
-      .click();
-    await page.getByRole('option', { name: 'http', exact: true }).click();
-
-    await activeSection.getByLabel('Timeout', { exact: true }).fill('5');
-    await activeSection.getByLabel('Concurrency', { exact: true }).fill('2');
-    await activeSection
-      .getByLabel('Host', { exact: true })
-      .fill('health.example.com');
-    await activeSection.getByLabel('Port', { exact: true }).fill('8888');
-    await activeSection
-      .getByLabel('HTTP Path', { exact: true })
-      .fill('/health');
-
-    // Set the Unhealthy part of Active health check settings
-    const activeUnhealthySection = activeSection.getByRole('group', {
-      name: 'Unhealthy',
-    });
-    await activeUnhealthySection.getByLabel('Interval').fill('1');
-    await activeUnhealthySection.getByLabel('HTTP Failures').fill('3');
-    await activeUnhealthySection.getByLabel('TCP Failures').fill('3');
-    await activeUnhealthySection.getByLabel('Timeouts').fill('3');
-    await fillHTTPStatuses(
-      activeUnhealthySection.getByLabel('HTTP Statuses'),
-      '429',
-      '500',
-      '503'
-    );
-
-    // Activate passive health check
-    await healthCheckSection
-      .getByTestId('checksPassiveEnabled')
-      .locator('..')
-      .click();
-
-    // Set the Healthy part of Passive health check settings
-    const passiveSection = healthCheckSection.getByRole('group', {
-      name: 'Passive',
-    });
-    await passiveSection
-      .getByRole('textbox', { name: 'Type', exact: true })
-      .click();
-    await page.getByRole('option', { name: 'http', exact: true }).click();
-
-    // Set the Unhealthy part of Passive health check settings
-    const passiveUnhealthySection = passiveSection.getByRole('group', {
-      name: 'Unhealthy',
-    });
-    await passiveUnhealthySection.getByLabel('HTTP Failures').fill('3');
-    await passiveUnhealthySection.getByLabel('TCP Failures').fill('3');
-    await passiveUnhealthySection.getByLabel('Timeouts').fill('3');
-    await fillHTTPStatuses(
-      passiveUnhealthySection.getByLabel('HTTP Statuses'),
-      '500'
-    );
+  await uiFillUpstreamAllFields(test, page, {
+    name: upstreamNameWithAllFields,
+    desc: description,
   });
 
   // Submit the form
@@ -278,215 +64,10 @@ test('should CRUD upstream with all fields', async ({ page }) => {
   await upstreamsPom.isDetailPage(page);
 
   await test.step('verify all fields in detail page', async () => {
-    // Verify basic information
-    const name = page.getByLabel('Name', { exact: true });
-    await expect(name).toHaveValue(upstreamNameWithAllFields);
-    await expect(name).toBeDisabled();
-
-    const descriptionField = page.getByLabel('Description');
-    await expect(descriptionField).toHaveValue(description);
-    await expect(descriptionField).toBeDisabled();
-
-    // Verify node information
-    const nodesSection = page.getByRole('group', { name: 'Nodes' });
-    await expect(
-      nodesSection.getByRole('cell', { name: 'node1.example.com' })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '8080' })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '10', exact: true })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '1', exact: true })
-    ).toBeVisible();
-
-    await expect(
-      nodesSection.getByRole('cell', { name: 'node2.example.com' })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '8081' })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '5', exact: true })
-    ).toBeVisible();
-    await expect(
-      nodesSection.getByRole('cell', { name: '2', exact: true })
-    ).toBeVisible();
-
-    // Verify load balancing type
-    const loadBalancingSection = page.getByRole('group', {
-      name: 'Load Balancing',
+    await uiCheckUpstreamAllFields(page, {
+      name: upstreamNameWithAllFields,
+      desc: description,
     });
-    const typeField = loadBalancingSection.getByRole('textbox', {
-      name: 'Type',
-      exact: true,
-    });
-    await expect(typeField).toHaveValue('chash');
-    await expect(typeField).toBeDisabled();
-
-    // Verify Hash On field
-    const hashOnField = loadBalancingSection.getByRole('textbox', {
-      name: 'Hash On',
-      exact: true,
-    });
-    await expect(hashOnField).toHaveValue('header');
-    await expect(hashOnField).toBeDisabled();
-
-    // Verify Key field
-    const keyField = loadBalancingSection.getByLabel('Key');
-    await expect(keyField).toHaveValue('X-Custom-Header');
-    await expect(keyField).toBeDisabled();
-
-    // Verify protocol (Scheme)
-    const schemeField = page.getByRole('textbox', {
-      name: 'Scheme',
-      exact: true,
-    });
-    await expect(schemeField).toHaveValue('https');
-    await expect(schemeField).toBeDisabled();
-
-    // Verify retry count field (Retries)
-    const retriesField = page.getByLabel('Retries');
-    await expect(retriesField).toHaveValue('5');
-    await expect(retriesField).toBeDisabled();
-
-    // Verify retry timeout field (Retry Timeout)
-    const retryTimeoutField = page.getByLabel('Retry Timeout');
-    await expect(retryTimeoutField).toHaveValue('6s');
-    await expect(retryTimeoutField).toBeDisabled();
-
-    // Verify Pass Host field
-    const passHostSection = page.getByRole('group', { name: 'Pass Host' });
-    const passHostField = passHostSection.getByRole('textbox', {
-      name: 'Pass Host',
-      exact: true,
-    });
-    await expect(passHostField).toHaveValue('rewrite');
-    await expect(passHostField).toBeDisabled();
-
-    // Verify Upstream Host field
-    const upstreamHostField = page.getByLabel('Upstream Host');
-    await expect(upstreamHostField).toHaveValue('custom.upstream.host');
-    await expect(upstreamHostField).toBeDisabled();
-
-    // Verify timeout settings (Timeout)
-    const timeoutSection = page.getByRole('group', { name: 'Timeout' });
-    await expect(timeoutSection.getByLabel('Connect')).toHaveValue('3s');
-    await expect(timeoutSection.getByLabel('Send')).toHaveValue('3s');
-    await expect(timeoutSection.getByLabel('Read')).toHaveValue('3s');
-
-    // Verify keepalive pool settings (Keepalive Pool)
-    const keepaliveSection = page.getByRole('group', {
-      name: 'Keepalive Pool',
-    });
-    await expect(keepaliveSection.getByLabel('Size')).toHaveValue('320');
-    await expect(keepaliveSection.getByLabel('Idle Timeout')).toHaveValue(
-      '60s'
-    );
-    await expect(keepaliveSection.getByLabel('Requests')).toHaveValue('1000');
-
-    // Verify TLS settings
-    const tlsSection = page.getByRole('group', { name: 'TLS' });
-    await expect(tlsSection.getByLabel('Verify')).toBeChecked();
-
-    // Verify health check settings
-    const healthCheckSection = page.getByRole('group', {
-      name: 'Health Check',
-    });
-    // Check if Active and Passive health checks are enabled (by checking if the respective sections exist)
-    await expect(
-      healthCheckSection.getByRole('group', { name: 'Active' })
-    ).toBeVisible();
-    await expect(
-      healthCheckSection.getByRole('group', { name: 'Passive' })
-    ).toBeVisible();
-
-    // Verify active health check settings
-    const activeSection = healthCheckSection.getByRole('group', {
-      name: 'Active',
-    });
-    const activeTypeField = activeSection.getByRole('textbox', {
-      name: 'Type',
-      exact: true,
-    });
-    await expect(activeTypeField).toHaveValue('http');
-    // Use more specific selectors for Timeout to avoid ambiguity
-    await expect(
-      activeSection.getByRole('textbox', { name: 'Timeout', exact: true })
-    ).toHaveValue('5s');
-    await expect(activeSection.getByLabel('Concurrency')).toHaveValue('2');
-    await expect(activeSection.getByLabel('Host')).toHaveValue(
-      'health.example.com'
-    );
-    await expect(activeSection.getByLabel('Port')).toHaveValue('8888');
-    await expect(activeSection.getByLabel('HTTP Path')).toHaveValue('/health');
-
-    // Verify passive health check settings
-    const passiveSection = healthCheckSection.getByRole('group', {
-      name: 'Passive',
-    });
-
-    // Verify active health check - healthy status settings
-    const activeHealthySection = activeSection.getByRole('group', {
-      name: 'Healthy',
-    });
-    // Check if the Successes field exists rather than its exact value
-    // This is more resilient to UI differences
-    await expect(activeHealthySection.getByLabel('Successes')).toBeVisible();
-
-    // Verify active health check - unhealthy status settings
-    const activeUnhealthySection = activeSection.getByRole('group', {
-      name: 'Unhealthy',
-    });
-    // Check if the fields exist rather than their exact values
-    // This is more resilient to UI differences
-    await expect(
-      activeUnhealthySection.getByLabel('HTTP Failures')
-    ).toBeVisible();
-    await expect(
-      activeUnhealthySection.getByLabel('TCP Failures')
-    ).toBeVisible();
-    await expect(activeUnhealthySection.getByLabel('Timeouts')).toBeVisible();
-    // Skip HTTP Statuses verification since the format might be different in detail view
-
-    // Verify passive health check settings
-    const passiveTypeField = passiveSection.getByRole('textbox', {
-      name: 'Type',
-      exact: true,
-    });
-    // Check if the Type field exists and is visible
-    await expect(passiveTypeField).toBeVisible();
-
-    // Verify passive health check - healthy status settings
-    const passiveHealthySection = passiveSection.getByRole('group', {
-      name: 'Healthy',
-    });
-    // Check if the Successes field exists rather than its exact value
-    await expect(passiveHealthySection.getByLabel('Successes')).toBeVisible();
-
-    // Verify passive health check - unhealthy status settings
-    const passiveUnhealthySection = passiveSection.getByRole('group', {
-      name: 'Unhealthy',
-    });
-    // Check if the fields exist rather than their exact values
-    await expect(
-      passiveUnhealthySection.getByLabel('HTTP Failures')
-    ).toBeVisible();
-    await expect(
-      passiveUnhealthySection.getByLabel('TCP Failures')
-    ).toBeVisible();
-    await expect(passiveUnhealthySection.getByLabel('Timeouts')).toBeVisible();
-
-    // Verify that the HTTP Statuses section exists in some form
-    // We'll use a more general selector that should work regardless of the exact UI structure
-    await expect(
-      passiveSection.getByRole('group', { name: 'Unhealthy' })
-    ).toBeVisible();
-
-    // Note: We're not checking for specific HTTP status codes like '500'
-    // as the format might be different in the detail view
   });
 
   await test.step('return to list page and verify', async () => {
